requirements.txt
pymongo~=4.6.3
beautifulsoup4~=4.13.4
cryptography~=44.0.2
dotenv~=0.9.9
python-dotenv~=1.1.0
selenium~=4.31.0
python-telegram-bot~=22.0